+++ sdks/outpost-typescript/RELEASES.md
@@ -89,3 +89,13 @@ Based on:
 - [typescript v0.9.1] sdks/outpost-typescript
 ### Releases
 - [NPM v0.9.1] https://www.npmjs.com/package/@hookdeck/outpost-sdk/v/0.9.1 - sdks/outpost-typescript
+
+## 2026-03-24 17:06:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.1 (2.869.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.9.2] sdks/outpost-typescript
+### Releases
+- [NPM v0.9.2] https://www.npmjs.com/package/@hookdeck/outpost-sdk/v/0.9.2 - sdks/outpost-typescript